Move Attribute and Tag to header files
