+++ AttributeParser/AttributeParser.vcxproj
@@ -154,7 +154,12 @@
     <ClCompile Include="AttributeParser.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <Text Include="attributes_test_0.txt" />
     <Text Include="tags_test_0.txt" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Attribute.h" />
+    <ClInclude Include="Tag.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
