+++ AttributeParser/attributes_test_0.txt
@@ -1,0 +1,14 @@
+<tag1  name = "hello">
+	<tag2>
+		<tag3>
+			<tag4>
+			</tag4>
+		</tag3>
+		<tag3.2>
+			<tag4.2>
+			</tag4.2>
+		</tag3.2>
+	</tag2>
+	<tag2.2>
+	</tag2.2>
+</tag1>
